+++ application/db/migrations/20180930172920_create_napicer_skills_table.rb
@@ -2,15 +2,16 @@ Sequel.migration do
   up do
     create_table(:napicer_skill) do
       primary_key :id
-      Integer :skill_id, null: false
+      foreign_key :skill_id, :skill, null: false
       String :napicer_id, null: false
       String :level, null: false
+      String :years_of_experience, null: false
       DateTime :created_at
       DateTime :updated_at
     end
   end
 
   down do
-    drop_table(:napicer)
+    drop_table(:napicer_skill)
   end
 end