+++ application/db/migrations/20181010161204_add_demo_table.rb
@@ -1,16 +1,19 @@
 Sequel.migration do
   up do
-    create_table(:napicer) do
+    create_table(:demo) do
       primary_key :id
       String :first_name, null: false
       String :last_name, null: false
-      String :username, null: false
+      String :email, null: false
+      String :phone, null: false
+      String :company_name, null: false
+      String :company_website, null: false
       DateTime :created_at
       DateTime :updated_at
     end
   end
 
   down do
-    drop_table(:napicer)
+    drop_table(:demo)
   end
 end