+++ spec/operations/user/create_spec.rb
@@ -1,11 +1,11 @@
 RSpec.describe Api::Operations::User::Create do
-  let(:args) do
+  let(:user) do
     {
       'email' => Faker::Internet.email
     }
   end
 
-  subject { described_class.new(args) }
+  subject { described_class.new(user: user) }
 
   it 'adds record successfully' do
     expect { subject.create }.to(change { Api::Models::User.count })